Day 14: Scope
Objective
The absolute difference between two integers,  and, is written as. The maximum absolute difference between two integers in a set of positive integers is the largest absolute difference between any two integers in.

Task
Complete the Difference class by writing the following:
A class constructor takes an array of integers as a parameter and saves it to the instance variable.
A computeDifference method that finds the maximum absolute difference between any numbers in and stores it in the instance variable.